% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_dat.R
\name{save_dat}
\alias{save_dat}
\title{Process and saves data of give user input (name and path)}
\usage{
save_dat(df, position, out_path)
}
\arguments{
\item{df}{a dataframe that contains columns \code{pos} and \code{overall}}

\item{position}{a string that filters the players of the df}

\item{out_path}{a string path(absolute) that tells where to save data}
}
\value{
a csv file that has name of \code{out_path/position.csv} format.
}
\description{
Process and saves data of give user input (name and path)
}
\examples{
save_dat(data.frame(pos = c("A", "B"), overall = c(1,2)),"A", "../temp/")
save_dat(data.frame(pos = c("A","B"), overall = c(0,0)), "", "../temp/")
}
